Move faucet request handling to faucetfunds.go
